Update TL Prizes activity workflow
