+++ app/components/ActivityList/index.js
@@ -13,7 +13,6 @@ import { getLastScheduled, getNextScheduled, getScheduledNotifications } from '.
 import { delayedExec, clearExec } from '../../services/timing';
 import sortActivities from './sortActivities';
 import ActivityListItem from './ActivityListItem';
-import TokenPrizesActivity from '../TokenPrizesActivity';
 import {
   newAppletSelector,
   activitySelectionDisabledSelector,
@@ -27,159 +26,7 @@ import {
   inProgressSelector,
 } from '../../state/responses/responses.selectors';
 
-import { prizesActivity } from './prizesActivity';
-
-const dateParser = (schedule) => {
-  const output = {};
-  Object.keys(schedule.events).forEach(key => {
-    const e = schedule.events[key];
-    const uri = e.data.URI;
-
-    if (!output[uri]) {
-      output[uri] = {
-        notificationDateTimes: [],
-        invalid: e.valid,
-      };
-    }
-
-    const eventSchedule = Parse.schedule(e.schedule);
-    const now = Day.fromDate(new Date());
-
-    const lastScheduled = getLastScheduled(eventSchedule, now);
-    const nextScheduled = getNextScheduled(eventSchedule, now);
-
-    const notifications = R.pathOr([], ['data', 'notifications'], e);
-    const dateTimes = getScheduledNotifications(eventSchedule, now, notifications);
-
-    let lastScheduledResponse = lastScheduled;
-    let {
-      lastScheduledTimeout, lastTimedActivity, extendedTime, invalid, completion
-    } = output[uri];
-
-    if (lastScheduledResponse) {
-      lastScheduledTimeout = e.data.timeout;
-      lastTimedActivity = e.data.timedActivity;
-      completion = e.data.completion;
-      invalid = e.valid;
-      extendedTime = e.data.extendedTime;
-    }
-
-    if (output[uri].lastScheduledResponse && lastScheduled) {
-      lastScheduledResponse = moment.max(
-        moment(output[uri].lastScheduledResponse),
-        moment(lastScheduled),
-      );
-      if (lastScheduledResponse === output[uri].lastScheduledResponse) {
-        lastScheduledTimeout = output[uri].lastScheduledTimeout;
-        lastTimedActivity = output[uri].lastTimedActivity;
-        invalid = output[uri].valid;
-        completion = output[uri].completion;
-        extendedTime = output[uri].extendedTime;
-      }
-    }
-
-    let nextScheduledResponse = nextScheduled;
-    let { nextScheduledTimeout, nextTimedActivity } = output[uri];
-
-    if (nextScheduledResponse) {
-      nextScheduledTimeout = e.data.timeout;
-      nextTimedActivity = e.data.timedActivity;
-    }
-
-    if (output[uri].nextScheduledResponse && nextScheduled) {
-      nextScheduledResponse = moment.min(
-        moment(output[uri].nextScheduledResponse),
-        moment(nextScheduled),
-      );
-      if (nextScheduledResponse === output[uri].nextScheduledResponse) {
-        nextScheduledTimeout = output[uri].nextScheduledTimeout;
-        nextTimedActivity = output[uri].nextTimedActivity;
-      }
-    }
-
-    output[uri] = {
-      lastScheduledResponse: lastScheduledResponse || output[uri].lastScheduledResponse,
-      nextScheduledResponse: nextScheduledResponse || output[uri].nextScheduledResponse,
-      lastTimedActivity,
-      nextTimedActivity,
-      extendedTime,
-      invalid,
-      lastScheduledTimeout,
-      nextScheduledTimeout,
-      completion,
-      // TODO: only append unique datetimes when multiple events scheduled for same activity/URI
-      notificationDateTimes: output[uri].notificationDateTimes.concat(dateTimes),
-    };
-  });
-
-  return output;
-};
-
-const getActivities = (applet, responseSchedule) => {
-  let scheduledDateTimesByActivity = {};
-  // applet.schedule, if defined, has an events key.
-  // events is a list of objects.
-  // the events[idx].data.URI points to the specific activity's schema.
-  if (applet.schedule) {
-    scheduledDateTimesByActivity = dateParser(applet.schedule);
-  }
-
-  const extraInfoActivities = applet.activities.map((act) => {
-    const scheduledDateTimes = scheduledDateTimesByActivity[act.schema];
-    const nextScheduled = R.pathOr(null, ['nextScheduledResponse'], scheduledDateTimes);
-    const lastScheduled = R.pathOr(null, ['lastScheduledResponse'], scheduledDateTimes);
-    const nextTimedActivity = R.pathOr(null, ['nextTimedActivity'], scheduledDateTimes);
-    const lastTimedActivity = R.pathOr(null, ['lastTimedActivity'], scheduledDateTimes);
-    const oneTimeCompletion = R.pathOr(null, ['completion'], scheduledDateTimes);
-    const lastTimeout = R.pathOr(null, ['lastScheduledTimeout'], scheduledDateTimes);
-    const nextTimeout = R.pathOr(null, ['nextScheduledTimeout'], scheduledDateTimes);
-    const invalid = R.pathOr(null, ['invalid'], scheduledDateTimes);
-    const extendedTime = R.pathOr(null, ['extendedTime'], scheduledDateTimes);
-
-    const lastResponse = R.path([applet.id, act.id, 'lastResponse'], responseSchedule);
-    let nextAccess = false;
-    let prevTimeout = null;
-    let scheduledTimeout = null;
-
-    if (lastTimeout) {
-      prevTimeout = ((lastTimeout.day * 24 + lastTimeout.hour) * 60 + lastTimeout.minute) * 60000;
-    }
-    if (nextTimeout) {
-      nextAccess = nextTimeout.access;
-      scheduledTimeout = ((nextTimeout.day * 24 + nextTimeout.hour) * 60 + nextTimeout.minute) * 60000;
-    }
-
-    return {
-      ...act,
-      appletId: applet.id,
-      appletShortName: applet.name,
-      appletName: applet.name,
-      appletSchema: applet.schema,
-      appletSchemaVersion: applet.schemaVersion,
-      lastScheduledTimestamp: lastScheduled,
-      lastResponseTimestamp: lastResponse,
-      nextScheduledTimestamp: nextScheduled,
-      oneTimeCompletion: oneTimeCompletion || false,
-      lastTimeout: prevTimeout,
-      nextTimeout: scheduledTimeout,
-      nextTimedActivity,
-      lastTimedActivity,
-      currentTime: new Date().getTime(),
-      invalid,
-      extendedTime,
-      nextAccess,
-      isOverdue: lastScheduled && moment(lastResponse) < moment(lastScheduled),
-
-      // also add in our parsed notifications...
-      notification: R.prop('notificationDateTimes', scheduledDateTimes),
-    };
-  });
-
-  return {
-    ...applet,
-    activities: extraInfoActivities,
-  };
-};
+import { parseAppletActivities } from '../../models/json-ld';
 
 const ActivityList = ({
   applet,
@@ -198,20 +45,26 @@ const ActivityList = ({
   inProgress,
   onPressActivity,
   onLongPressActivity,
-  onPressPrize,
-  onLongPressPrize,
 }) => {
   // const newApplet = getActivities(applet.applet, responseSchedule);
   const [activities, setActivities] = useState([]);
+  const [prizeActivity, setPrizeActivity] = useState(null);
   const [isConnected, setIsConnected] = useState(true);
 
   const updateStatusDelay = 60 * 1000;
   const updateScheduleDelay = 24 * 3600 * 1000;
 
   const stateUpdate = () => {
-    const newApplet = getActivities(applet, responseSchedule);
+    const newApplet = parseAppletActivities(applet, responseSchedule);
 
-    setActivities(sortActivities(applet.id, newApplet.activities, inProgress, activityEndTimes));
+    const appletActivities = newApplet.activities.filter(act => act.isPrize != true);
+
+    setActivities(sortActivities(applet.id, appletActivities, inProgress, activityEndTimes));
+
+    const pzActs = newApplet.activities.filter(act => act.isPrize === true)
+    if (pzActs.length === 1) {
+      setPrizeActivity(pzActs[0]);
+    }
   };
 
   const datesAreOnSameDay = (first, second) => first.getFullYear() === second.getFullYear()
@@ -335,11 +188,13 @@ const ActivityList = ({
           key={activity.id || activity.text}
         />
       ))}
-      <TokenPrizesActivity
-        prizesActivity={prizesActivity}
-        onPress={onPressPrize}
-        onLongPress={onLongPressPrize}
-      />
+      {prizeActivity && (
+        <ActivityListItem
+          onPress={() => onPressActivity(prizeActivity)}
+          onLongPress={() => onLongPressActivity(prizeActivity)}
+          activity={prizeActivity}
+        />
+      )}
     </View>
   );
 };
@@ -353,8 +208,6 @@ ActivityList.propTypes = {
   inProgress: PropTypes.object.isRequired,
   onPressActivity: PropTypes.func.isRequired,
   onLongPressActivity: PropTypes.func.isRequired,
-  onPressPrize: PropTypes.func.isRequired,
-  onLongPressPrize: PropTypes.func.isRequired,
   lastUpdatedTime: PropTypes.object.isRequired,
   activityEndTimes: PropTypes.object.isRequired,
   setUpdatedTime: PropTypes.func.isRequired,
