+++ app/i18n/translations/en.json
@@ -298,7 +298,9 @@
     "hi": "Hi",
     "enter_old_password": "Please enter your old password",
     "enter_new_password": "Please enter a new password"
+  },
+  "prize": {
+    "lack_tokens_title": "Try again",
+    "lack_tokens_subtitle": "You don't have enough tokens.\nPlease select another prize"
   }
-
-
 }
