+++ app/scenes/AppletDetails/index.js
@@ -41,11 +41,6 @@ class AppletDetails extends Component {
     this.props.startResponse(activity);
   };
 
-  handlePressPrize = (prizesActivity) => {
-    this.props.setCurrentPrizesActivity(prizesActivity);
-    Actions.push('take_prizes');
-  }
-
   /**
    * Method called when the activity is pressed for a few seconds.
    *
@@ -119,8 +114,6 @@ class AppletDetails extends Component {
         onPressDrawer={Actions.drawerOpen}
         onPressActivity={this.handlePressActivity}
         onLongPressActivity={this.handleLongPressActivity}
-        onPressPrize={this.handlePressPrize}
-        onLongPressPrize={this.handleLongPressActivity}
         onPressBack={this.handleBack}
         onPressSettings={() => Actions.push("applet_settings")}
         primaryColor={skin.colors.primary}
@@ -149,7 +142,6 @@ AppletDetails.propTypes = {
   getAppletResponseData: PropTypes.func.isRequired,
   setAppletSelectionDisabled: PropTypes.func.isRequired,
   setActivitySelectionDisabled: PropTypes.func.isRequired,
-  setCurrentPrizesActivity: PropTypes.func.isRequired,
 };
 
 const mapStateToProps = (state) => ({
@@ -169,7 +161,6 @@ const mapDispatchToProps = {
   getAppletResponseData,
   setAppletSelectionDisabled,
   setActivitySelectionDisabled,
-  setCurrentPrizesActivity,
 };
 
 export default connect(
