+++ app/scenes/Login/index.js
@@ -23,6 +23,7 @@ import LoginForm from "./LoginForm";
 import {
   skinSelector,
   mobileDataAllowedSelector,
+  apiHostSelector,
 } from "../../state/app/app.selectors";
 import { toggleMobileDataAllowed } from "../../state/app/app.actions";
 
@@ -113,7 +114,7 @@ class Login extends Component {
   };
 
   render() {
-    const { skin, mobileDataAllowed, toggleMobileDataAllowed } = this.props;
+    const { skin, mobileDataAllowed, toggleMobileDataAllowed, apiHost } = this.props;
     const title = skin.name;
     const logo =
       typeof skin.logo !== "undefined" ? { uri: skin.logo } : defaultLogo;
@@ -190,6 +191,7 @@ const mapStateToProps = (state) => ({
   skin: skinSelector(state),
   mobileDataAllowed: mobileDataAllowedSelector(state),
   fcmToken: fcmFcmTokenSelector(state),
+  apiHost: apiHostSelector(state),
 });
 
 const mapDispatchToProps = {
@@ -199,5 +201,5 @@ const mapDispatchToProps = {
 
 export default connect(
   mapStateToProps,
-  mapDispatchToProps
+  mapDispatchToProps,
 )(Login);
