+++ app/state/app/app.constants.js
@@ -8,7 +8,6 @@ export default {
   SET_APP_STATUS: 'SET_APP_STATUS',
   SET_CURRENT_APPLET: 'SET_CURRENT_APPLET',
   SET_CURRENT_ACTIVITY: 'SET_CURRENT_ACTIVITY',
-  SET_CURRENT_PRIZES_ACTIVITY: 'SET_CURRENT_PRIZES_ACTIVITY',
   TOGGLE_MOBILE_DATA_ALLOWED: 'TOGGLE_MOBILE_DATA_ALLOWED',
   SET_APP_LANGUAGE: 'SET_APP_LANGUAGE',
   SET_ACTIVITY_START_TIME: 'SET_ACTIVITY_START_TIME',
