+++ app/state/app/app.reducer.js
@@ -35,13 +35,6 @@ export const initialState = {
   currentActivity: null,
 
   /**
-   * Current prizes activity.
-   *
-   * @type {string}
-   */
-  currentPrizesActivity: null,
-
-  /**
    * Whether the applet cards are disabled.
    *
    * @type {boolean}.
@@ -152,11 +145,6 @@ export default (state = initialState, action = {}) => {
       return {
         ...state,
         currentActivity: action.payload,
-      };
-    case APP_CONSTANTS.SET_CURRENT_PRIZES_ACTIVITY:
-      return {
-        ...state,
-        currentPrizesActivity: action.payload,
       };
     case APP_CONSTANTS.SET_APPLET_SELECTION_DISABLED:
       return {
