+++ app/state/user/user.constants.js
@@ -2,5 +2,6 @@ export default {
   SET_RESPONSE_COLLECTION_ID: 'SET_RESPONSE_COLLECTION_ID',
   SET_AUTH: 'SET_USER_AUTH',
   SET_INFO: 'SET_USER_INFO',
+  SET_TOKEN_BALANCE: 'SET_TOKEN_BALANCE',
   CLEAR: 'CLEAR_USER',
 };
