+++ app/state/user/user.reducer.js
@@ -3,6 +3,7 @@ import USER_CONSTANTS from './user.constants';
 export const initialState = {
   auth: null,
   info: null,
+  tokenBalance: 0,
 };
 
 export default (state = initialState, action = {}) => {
@@ -18,6 +19,11 @@ export default (state = initialState, action = {}) => {
       return {
         ...state,
         info: action.payload,
+      };
+    case USER_CONSTANTS.SET_TOKEN_BALANCE:
+      return {
+        ...state,
+        tokenBalance: action.payload,
       };
     default:
       return state;
