+++ app/state/user/user.selectors.js
@@ -6,4 +6,6 @@ export const authTokenSelector = R.path(['user', 'auth', 'token']);
 
 export const userInfoSelector = R.path(['user', 'info']);
 
+export const userTokenBalanceSelector = R.path(['user', 'tokenBalance']);
+
 export const loggedInSelector = R.pathOr(false, ['user', 'auth', 'token']);
